BC-334: Introduced tests for all assessment outcomes for crime invc, crime non-invc, and civil.
