+++ src/e2e/main/java/uk/gov/justice/laa/amend/claim/pages/ReviewAndAmendPage.java
@@ -89,7 +89,18 @@ public class ReviewAndAmendPage {
             .click();
     }
 
+    private void clickAddLink(String itemName, String id) {
+        Locator link = rowByItemName(itemName).getByTestId(id);
+        clickAddLink(link);
+    }
+
+    private void clickAddLink(Locator link) {
+        assertThat(link).containsText("Add");
+        link.click();
+    }
+
     public void clickChangeProfitCosts()      { clickAddChangeRow("Profit costs"); }
+    public void clickAddProfitCosts()         { clickAddLink("Profit costs", "claim-field-profitCost"); }
     public void clickChangeDisbursements()    { clickChangeInRow("Disbursements"); }
     public void clickChangeDisbursementsVat() { clickChangeInRow("Disbursement VAT"); }
     public void clickChangeTravelCosts()      { clickChangeInRow("Travel costs"); }
@@ -110,22 +121,14 @@ public class ReviewAndAmendPage {
 
 
     public void clickAddAssessedTotalVat() {
-        totalClaimValueTable.getByTestId("totals-assessedTotalVat").click();
+        Locator link = totalClaimValueTable.getByTestId("totals-assessedTotalVat");
+        clickAddLink(link);
     }
-
-    public void clickAddAssessedTotalInclVat() {
-        totalClaimValueTable.getByTestId("totals-assessedTotalInclVat").click();
-    }
-
 
     public void clickAddAllowedTotalVat() {
-        totalAllowedValueTable.getByTestId("totals-allowedTotalVat").click();
+        Locator link = totalAllowedValueTable.getByTestId("totals-allowedTotalVat");
+        clickAddLink(link);
     }
-
-    public void clickAddAllowedTotalInclVat() {
-        totalAllowedValueTable.getByTestId("totals-allowedTotalInclVat").click();
-    }
-
 
     public void submitAdjustments() {
         submitAdjustmentsButton.click();
