+++ src/e2e/test/java/uk/gov/justice/laa/amend/claim/tests/DiscardAssessmentTest.java
@@ -64,7 +64,6 @@ public class DiscardAssessmentTest extends BaseTest {
                 .id(CLAIM_ID)
                 .submissionId(SUBMISSION_ID)
                 .uniqueFileNumber(UFN)
-                .feeCode("INVC")
                 .userId(USER_ID)
                 .build(),
